requirements.txt
streamlit
pandas
numpy
scikit-learn
torch
plotly
python-nmap
scapy
langchain
python-dotenv
langchain-google-genai
google-generativeai